Move io-related files into io folder
