+++ include/ozo/impl/async_request.h
@@ -4,9 +4,9 @@
 #include <ozo/detail/post_handler.h>
 #include <ozo/detail/timeout_handler.h>
 #include <ozo/impl/io.h>
-#include <ozo/detail/array.h>
-#include <ozo/recv.h>
-#include <ozo/binary_query.h>
+#include <ozo/io/array.h>
+#include <ozo/io/recv.h>
+#include <ozo/io/binary_query.h>
 #include <ozo/connection.h>
 #include <ozo/query_builder.h>
 #include <ozo/time_traits.h>
