+++ include/ozo/io/istream.h
