+++ include/ozo/io/ostream.h
