+++ tests/binary_serialization.cpp
@@ -1,5 +1,5 @@
-#include <ozo/send.h>
-#include <ozo/detail/array.h>
+#include <ozo/io/send.h>
+#include <ozo/io/array.h>
 #include <ozo/ext/std.h>
 
 #include <gtest/gtest.h>
