+++ tests/result_integration.cpp
@@ -1,8 +1,8 @@
 #include <ozo/result.h>
-#include <ozo/recv.h>
+#include <ozo/io/recv.h>
 #include <ozo/ext/boost.h>
 #include <ozo/ext/std.h>
-#include <ozo/detail/array.h>
+#include <ozo/io/array.h>
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
